Move plot_graph out of common
